refactor(auth): correct import path
